[OPENENGSB-1006] renamed KeyValuePair, added a license header to the new file, some checkstyle warnings eliminated
